+++ voidstar/algos_3d/cube.h
@@ -7,24 +7,24 @@ class Algo3DCube : public Algo3D {
   virtual ~Algo3DCube() {}
 
   virtual bool apply(Floats& vertices, Floats& colors, VertIndices& selected,
-                     size_t width, size_t height, size_t depth) = 0;
+                     u32 width, u32 height, u32 depth) = 0;
 
  protected:
-  size_t make_vertices(Floats& vertices, size_t width, size_t height,
-                       size_t depth) {
+  u32 make_vertices(Floats& vertices, u32 width, u32 height,
+                       u32 depth) {
     const float w = static_cast<float>(width) / 2;
     const float h = static_cast<float>(height) / 2;
     const float d = static_cast<float>(depth) / 2;
 
     size_t pos = 0;
-    for (size_t z = 0; z < depth; ++z)
-      for (size_t y = 0; y < height; ++y)
-        for (size_t x = 0; x < width; ++x) {
+    for (u32 z = 0; z < depth; ++z)
+      for (u32 y = 0; y < height; ++y)
+        for (u32 x = 0; x < width; ++x) {
           vertices[pos++] = (static_cast<float>(x) - w) / w;
           vertices[pos++] = (static_cast<float>(y) - h) / h;
           vertices[pos++] = (static_cast<float>(z) - d) / d;
         }
 
-    return vertices.size() / 3;
+    return static_cast<u32>(vertices.size()) / 3;
   }
 };
